feat: Add kudos overview row component - MEED-2676 - Meeds-io/MIPs#81 (#403)
(cherry picked from commit 58553a0db763e099d30cd9ee768badf161c824bc)
