+++ kudos-webapps/src/main/webapp/vue-app/kudos-overview/components/KudosOverview.vue
@@ -4,12 +4,8 @@
     class="white">
     <widget-wrapper
       id="kudosOverviewHeader"
-      :default-padding="isOverviewDisplay && 'px-0' || 'px-5 pb-5'"
-      :hide-padding-top="isOverviewDisplay">
-      <template v-if="!isOverviewDisplay" #title>
-        <span class="widget-text-header text-truncate">{{ $t('exoplatform.kudos.button.rewardedKudos') }}</span>
-      </template>
-      <template v-if="!isOverviewDisplay" #action>
+      :title="$t('exoplatform.kudos.button.rewardedKudos')">
+      <template #action>
         <select
           v-model="periodType"
           class="kudosOverviewPeriodSelect fill-height col-auto my-auto py-0 subtitle-1 ignore-vuetify-classes">
@@ -21,39 +17,9 @@
           </option>
         </select>
       </template>
-      <v-row
-        id="kudosOverviewCardsParent"
-        class="white border-box-sizing px-4 py-0 ma-0 align-center">
-        <v-col class="kudosOverviewCard">
-          <kudos-overview-card
-            :clickable="owner && receivedKudosCount > 0"
-            class="kudosReceivedOverviewPeriod mx-n4"
-            @open-drawer="openDrawer('received')">
-            <template slot="count">
-              {{ receivedKudosCount || '0' }}
-            </template>
-            <template slot="label">
-              {{ $t('exoplatform.kudos.label.received') }}
-            </template> 
-          </kudos-overview-card>
-        </v-col>
-        <v-divider
-          class="my-9 mx-8 me-md-1 ms-md-5"
-          vertical />
-        <v-col class="kudosOverviewCard">
-          <kudos-overview-card
-            :clickable="owner && sentKudosCount > 0"
-            class="kudosSentOverviewPeriod mx-n4"
-            @open-drawer="openDrawer('sent')">
-            <template slot="count">
-              {{ sentKudosCount || '0' }}
-            </template>
-            <template slot="label">
-              {{ $t('exoplatform.kudos.label.sent') }}
-            </template>
-          </kudos-overview-card>
-        </v-col>
-      </v-row>
+      <kudos-overview-row 
+        :is-owner="owner" 
+        :period-type="periodType" />
     </widget-wrapper>
     <kudos-overview-drawer
       v-if="owner"
@@ -62,23 +28,12 @@
 </template>
 
 <script>
-import {getKudosSent, getKudosReceived} from '../../js/Kudos.js'; 
 
 export default {
-  props: {
-    isOverviewDisplay: {
-      type: Boolean,
-      default: () => false,
-    },
-  },
+
   data: () => ({
     owner: eXo.env.portal.profileOwner === eXo.env.portal.userName,
-    identityId: eXo.env.portal.profileOwnerIdentityId,
     periodType: 'WEEK',
-    sentKudosCount: 0,
-    receivedKudosCount: 0,
-    sentKudos: [],
-    receivedKudos: [],
   }),
   computed: {
     periods() {
@@ -100,44 +55,6 @@ export default {
       }];
     },
   },
-  watch: {
-    periodType() {
-      this.refresh();
-    },
-  },
-  created() {
-    this.refresh();
-  },
-  methods: {
-    openDrawer(kudosType) {
-      if (this.owner) {
-        const title = kudosType === 'sent' ?
-          this.$t('exoplatform.kudos.button.sentKudos'):
-          this.$t('exoplatform.kudos.button.receivedKudos');
-        this.$refs.kudosOverviewDrawer.open(title, kudosType, this.identityId, this.periodType);
-      }
-    },
-    refresh() {
-      const dateInSeconds = parseInt(Date.now() / 1000);
-      document.dispatchEvent(new CustomEvent('displayTopBarLoading'));
-      getKudosSent(this.identityId, 0, true, this.periodType, dateInSeconds)
-        .then(kudosList => {
-          this.sentKudosCount = kudosList && kudosList.size || 0;
-          this.sentKudos = kudosList && kudosList.kudos || [];
-        })
-        .then(() => getKudosReceived(this.identityId, 0, true, this.periodType, dateInSeconds))
-        .then(kudosList => {
-          this.receivedKudosCount = kudosList && kudosList.size || 0;
-          this.receivedKudos = kudosList && kudosList.kudos || [];
-          return this.$nextTick();
-        })
-        .finally(() => {
-          this.$root.$emit('application-loaded');
-          // Decrement 'loading' effect in top of the page
-          document.dispatchEvent(new CustomEvent('hideTopBarLoading'));
-          document.dispatchEvent(new CustomEvent('kudosCount', {detail: this.sentKudosCount + this.receivedKudosCount}));
-        });
-    },
-  },
+
 };
 </script>
